Change cavoke.h interface
