+++ games/cavoke.h
@@ -6,7 +6,9 @@
 #include <string>
 #include <vector>
 
+// TODO: logging
 namespace cavoke {
+using json = nlohmann::json;
 
 struct GameState {  // NOLINT(cppcoreguidelines-pro-type-member-init)
     bool is_terminal;
@@ -29,36 +31,30 @@ struct GameMove {  // NOLINT(cppcoreguidelines-pro-type-member-init)
 
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE(GameMove, player_id, move, global_state)
 
-GameState init_state();
+namespace {
+
+struct ValidationResult {
+    bool success;
+    std::string message;
+};
+
+NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE(ValidationResult, success, message);
+
+}  // namespace
+
+// called every time admin changes settings or new player arrives
+// returns true if game can be started with current settings; false otherwise
+// use message_callback(<some_message>) to display message to the players
+bool validate_settings(const json &settings, const std::vector<int> &occupied_positions, const std::function<void(std::string)> &message_callback);
+
+// called once when the game session is started
+// generates initial state
+// settings have passed the validation
+GameState init_state(const json &settings, const std::vector<int> &occupied_positions);
+
+// applies move to the state
 GameState apply_move(GameMove &new_move);
 
-// TODO: logging
-
 }  // namespace cavoke
-
-using json = nlohmann::json;
-
-int main() {
-    std::string input;
-    char c;
-    while (std::cin.get(c)) {
-        input.push_back(c);
-    }
-    std::cerr << "RECEIVED MOVE: '" << input << "'" << std::endl;
-    json input_json = json::parse(input);
-    cavoke::GameMove move = input_json.get<cavoke::GameMove>();
-
-    cavoke::GameState new_state;
-    if (move.player_id == -1) {
-        new_state = cavoke::init_state();
-    } else {
-        new_state = cavoke::apply_move(move);
-    }
-
-    json result_json = new_state;
-
-    std::cerr << "SEND RESULT: '" << result_json << "'" << std::endl;
-    std::cout << result_json;
-}
 
 #endif  // CAVOKE_CAVOKE_H
